Refactor: reorganize CLI abstractions and namespaces
