+++ src/Abstractions/Commands/FuncCommandInvocationContext.cs
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-namespace Azure.Functions.Cli.Workloads;
+namespace Azure.Functions.Cli.Commands;
 
 /// <summary>
 /// Surface a <see cref="FuncCommand"/> uses at execution time to read parsed option
