+++ src/Abstractions/Projects/IProjectResolver.cs
@@ -1,7 +1,9 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-namespace Azure.Functions.Cli.Workloads;
+using Azure.Functions.Cli.Workloads;
+
+namespace Azure.Functions.Cli.Projects;
 
 /// <summary>
 /// Decides whether a directory is a project owned by this workload.
