+++ src/Func/Commands/InitCommand.cs
@@ -5,6 +5,7 @@ using System.CommandLine;
 using Azure.Functions.Cli.Common;
 using Azure.Functions.Cli.Console;
 using Azure.Functions.Cli.Hosting;
+using Azure.Functions.Cli.Projects;
 using Azure.Functions.Cli.Workloads;
 
 namespace Azure.Functions.Cli.Commands;
