+++ src/Workload/DotNet/DotNetProjectInitializer.cs
@@ -2,7 +2,8 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System.CommandLine;
-using Azure.Functions.Cli.Workloads;
+using Azure.Functions.Cli.Commands;
+using Azure.Functions.Cli.Projects;
 
 namespace Azure.Functions.Cli.Workload.DotNet;
