+++ src/Workload/Node/NodeProjectResolver.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System.Text.Json;
-using Azure.Functions.Cli.Workloads;
+using Azure.Functions.Cli.Projects;
 
 namespace Azure.Functions.Cli.Workload.Node;
