+++ src/Workload/Python/PythonProjectResolver.cs
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using Azure.Functions.Cli.Workloads;
+using Azure.Functions.Cli.Projects;
 
 namespace Azure.Functions.Cli.Workload.Python;
