+++ test/Func.Tests/Workloads/ExternalCommandTests.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System.CommandLine;
+using Azure.Functions.Cli.Commands;
 using Azure.Functions.Cli.Workloads;
 using Xunit;
