+++ test/Func.Tests/Workloads/Resolution/WorkloadResolverTests.cs
@@ -1,6 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
+using Azure.Functions.Cli.Projects;
 using Azure.Functions.Cli.Workloads;
 using Azure.Functions.Cli.Workloads.Resolution;
 using NSubstitute;
