+++ test/Workload.Tests.Fixtures.WithCommand/StubWorkload.cs
@@ -1,6 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
+using Azure.Functions.Cli.Commands;
 using Azure.Functions.Cli.Workloads;
 
 namespace Azure.Functions.Cli.Workload.Tests.Fixtures.WithCommand;
