+++ test/Workload/Node.Tests/NodeProjectInitializerTests.cs
@@ -2,8 +2,8 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System.CommandLine;
+using Azure.Functions.Cli.Commands;
 using Azure.Functions.Cli.Common;
-using Azure.Functions.Cli.Workloads;
 using Xunit;
 
 namespace Azure.Functions.Cli.Workload.Node.Tests;
